requirements.txt
google-generativeai
requests
pytest

